결함 #179453 설문>설문작성>질문에파일첨부>질문복사>첨부파일삭제>복사대상이 된 질문의 첨부파일 깨짐
원인1. 질의 복사 > 원본,복사된 질의 파일삭제 하면 바로 서버에서 파일삭제 하여 발생한 이슈
원인2. 원본,복사된 질의가 같은 파일을 경로로 잡고있음

1. 복사 할때 파일도 새로운 파일객채로 생성 필요
2. 파일을 x버튼클릭하여 삭제시 바로 삭제 하지않고 수정 버튼 클릭시 삭제되도록 수정 필요
	객체에 삭재할 파일 경로 푸쉬
	질의 수정 완료시 객체 엘리먼트 생성 및
		질의 내부에서 취소하면 해당 객체의 id 값 배열 비우기
	최종 수정버튼 클릭시 do it 질의 아이디에 맞는 값들 초기화 하는 함수 작성

	+ 질의 복사하기 엘리먼트 에러 수정

+++ src/main/webapp/js/ezSurvey/surveyFile.js
@@ -302,37 +302,58 @@ var SurveyFile = function() {
 		function deleteFile(imgObj, event) {
 			if (event) {event.stopPropagation();}
 			
-			var liElmt   = imgObj.parentElement;
-			var filePath = liElmt.getAttribute("path");
-			
-			// 2020-03-13 김민성 - 질문 수정 > url 삭제시 url 삭제 메소드 분기 타도록 추가
-			if(filePath == null) {
-				deleteUrlFile(imgObj, event);
-				return;
+			var liElmt = imgObj.parentElement;
+			if (typeof(liElmt) != 'undefined' || liElmt != null ) {
+			    var setId = imgObj.closest('.qstnWrapper').id;
+			    if (!deleteFilePathObj[setId]) {
+                    deleteFilePathObj[setId] = [];
+                }
+			    deleteFilePathObj[setId].push(liElmt.getAttribute("path"));
+			    afterDeleteSuccessfully(liElmt);
 			}
-			
-			$.ajax({
-				type: "POST",
-				url: "/ezSurvey/deleteAttachFile.do",
-				data: {"filePath" : filePath},
-				dataType: "JSON",
-				async: false,
-				cache: false,
-				success : function(data) {
-					var code = data.code;
-					
-					switch(code) {
-						case 0 : afterDeleteSuccessfully(liElmt)  ; break;
-						case 1 : alert(SurveyMessages.strParamErr); break;
-						case 2 : alert(SurveyMessages.strError)   ; break;
-						default: alert(SurveyMessages.strError)   ; return;
-					}
-				},
-				error : function(error) {
-					alert(SurveyMessages.strError + error);
-				}
-			});
 		}
+		
+		function doDeleteFile() {
+            if (Object.keys(deleteFilePathObj).length === 0) {
+                return
+            }
+            Object.values(deleteFilePathObj).forEach(function(files) {
+            
+                files.forEach(function(v) {
+                    var filePath = v;
+                    if(filePath == null) {
+                        deleteUrlFile(imgObj, event);
+                        return;
+                    }
+                    
+                    $.ajax({
+                        type: "POST",
+                        url: "/ezSurvey/deleteAttachFile.do",
+                        data: {"filePath" : filePath},
+                        dataType: "JSON",
+                        async: false,
+                        cache: false,
+                        success : function(data) {
+                            /*
+                            var code = data.code;
+                            var liElmt = document.querySelector('[path="'+filePath+'"]')
+                            switch(code) {
+                                case 0 : afterDeleteSuccessfully(liElmt)  ; break;
+                                case 1 : alert(SurveyMessages.strParamErr); break;
+                                case 2 : alert(SurveyMessages.strError)   ; break;
+                                default: alert(SurveyMessages.strError)   ; return;
+                            }
+                            */
+                        },
+                        error : function(error) {
+                            alert(SurveyMessages.strError + error);
+                        }
+                    });
+                });
+            });
+            
+            deleteFilePathObj = {};
+        }
 		
 		function afterDeleteSuccessfully(liElmt) {
 			displayInputTitle(liElmt);
@@ -550,7 +571,8 @@ var SurveyFile = function() {
 			mkImgTag   : mkImgTag,
 			download   : downloadFile,
 			render     : renderAttachList,
-			makeImgTitle : makeImgTitle
+			makeImgTitle : makeImgTitle,
+			doDeleteFile : doDeleteFile
 		};
 	}
 }();
